+++ web_interface/delegate.go
@@ -3,5 +3,5 @@ package web_interface
 type WebDelegate interface {
   GetF(key string) (string,error)
   PutF(string,string,map[string]bool) (error,map[string]string)
-  Info(int) []uint
+  Info(int) []uint64
 }
